% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkParens.R
\name{checkParens2}
\alias{checkParens2}
\title{Determine if  parentheses in a given text are properly nested.}
\usage{
checkParens2(text, parens = matrix(c("(", "{", "[", ")", "}", "]"), ncol = 2))
}
\arguments{
\item{text}{character vector}
}
\value{
index of first invalid bracket or 0
}
\description{
Determine if  parentheses in a given text are properly nested.
}
\examples{
checkParens1("(a (( b ) ( c )) ( d ) e )")
checkParens2("(* (+ `(1 2 3)) 4)")
}

