% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csvToCsv2.R
\name{csvToCsv2}
\alias{csvToCsv2}
\title{Substitute field separators "," for ";"
and decimal separators "." for ","}
\usage{
csvToCsv2(infname, outfname)
}
\arguments{
\item{infname}{character, input path}

\item{outfname}{character, output path.}
}
\description{
Substitute field separators "," for ";"
and decimal separators "." for ","
}

