% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decdig.R
\name{decdig}
\alias{decdig}
\title{Create vector with n most significant decimal digits of the given number}
\usage{
decdig(x, n)
}
\arguments{
\item{x}{a single numeric value}

\item{n}{a single natural number, n <= 16}
}
\value{
integer vector with n most significant decimal digits of x
}
\description{
Create vector with n most significant decimal digits of the given number
}
\examples{
decdig(pi, 3)
decdig(exp(1), 5)
decdig(1e-2, 3)
}

