% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_prcomp.R
\name{extract_legend}
\alias{extract_legend}
\title{Extract label from ggplot
http://stackoverflow.com/a/13650878/1560062}
\usage{
extract_legend(p)
}
\arguments{
\item{p}{ggplot}
}
\value{
legend
}
\description{
Extract label from ggplot
http://stackoverflow.com/a/13650878/1560062
}

