% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fisher_yates_shuffle.R
\name{fisher_yates_shuffle}
\alias{fisher_yates_shuffle}
\title{Fisher–Yates shuffle}
\usage{
fisher_yates_shuffle(x)
}
\arguments{
\item{x}{atomic vector}
}
\value{
shuffled x
}
\description{
Fisher–Yates shuffle
}

