% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gendiscrete.R
\name{gendiscrete}
\alias{gendiscrete}
\title{Generate random sequence of elements choosen with replacement from given vetcor}
\usage{
gendiscrete(n, x, p)
}
\arguments{
\item{n}{single natural number}

\item{x}{numeric vector of length k with unique elements}

\item{p}{vector of probabilities}
}
\value{
vector of length n and type == type(x)
}
\description{
Generate random sequence of elements choosen with replacement from given vetcor
}

