% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/goldenratio.R
\name{golden_ratio}
\alias{golden_ratio}
\title{Determine a local minimum of a continuous real function f}
\usage{
golden_ratio(f, a, b, eps = 1e-16, maxiter = 100)
}
\arguments{
\item{f}{function}

\item{a}{numeric}

\item{b}{numeric > a}

\item{eps}{numeric (default 1e-16)}

\item{maxiter}{maximum number of iterations}
}
\value{
list containing approximation of local minimum,
  value at local minimum and some control data
}
\description{
Determine a local minimum of a continuous real function f
}
\examples{
golden_ratio(sin, -pi, pi)
golden_ratio(sign, -1, 1, eps = 1e-3)
golden_ratio(function(x) -1, -100, 100, maxiter = 10000)
}

