% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inqusition.R
\name{inquisition}
\alias{inquisition}
\title{For given adjacency matrix check if any two not directly connected vertices
have common following node}
\usage{
inquisition(graph)
}
\arguments{
\item{graph}{binary adjacency matrix}
}
\value{
single logical value
}
\description{
For given adjacency matrix check if any two not directly connected vertices
have common following node
}
\examples{
inquisition(matrix(ifelse(runif(100) > 0.75, 1, 0), nrow = 10))
}

