% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lcs}
\alias{lcs}
\title{Compute the length of the longest common subsequence of two numeric vectors.}
\usage{
lcs(first, second, trace = FALSE)
}
\arguments{
\item{first}{numeric vector}

\item{second}{numeric vector}
}
\value{
length of the lcs
}
\description{
Compute the length of the longest common subsequence of two numeric vectors.
}
\examples{
lcs(charToRaw("GAGAGTAGATAG"), charToRaw("ATA"))
lcs(charToRaw("FOOBAR"), charToRaw("BARFOO"))
}

