% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/logiderle.R
\name{logiderle}
\alias{logiderle}
\title{Retrun logical vector w of length n where
w_{l} == TRUE if there exists such p that l %in% i[p]:j[p]}
\usage{
logiderle(i, j, n)
}
\arguments{
\item{i}{an integer vector of lenght m}

\item{j}{an integer vector of lenght m}

\item{n}{natural number}
}
\value{
logical vector of length n
}
\description{
Retrun logical vector w of length n where
w_{l} == TRUE if there exists such p that l %in% i[p]:j[p]
}
\examples{
logiderle(c(1L, 4L), c(1L, 6L), 7L)
logiderle(c(1L), c(2L), 2L)
}

