% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\title{Determine the most frequently occurring value in an integer vector}
\usage{
mode(x)
}
\arguments{
\item{x}{integer vector}
}
\value{
integer mode of x
}
\description{
Determine the most frequently occurring value in an integer vector
}

