% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NAimput}
\alias{NAimput}
\title{Impute all missing values found in a given nondecreasingly ordered numeric vector x.}
\usage{
NAimput(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
copy of x with imputed NAs
}
\description{
Impute all missing values found in a given nondecreasingly ordered numeric vector x.
}
\examples{
NAimput(c(NA, NA, 4, 5, NA))
NAimput(c(1, NA, 2, NA, NA, 3))
}

