% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naomit}
\alias{naomit}
\title{Remove all missing values from a given numeric vector.}
\usage{
naomit(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Remove all missing values from a given numeric vector.
}

