% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/neighboreq.R
\name{neighboreq}
\alias{neighboreq}
\title{Given atomic vector x find all indices i where x[i] == x[i + 1]}
\usage{
neighboreq(x)
}
\arguments{
\item{x}{atomic vector}
}
\value{
vector of indices
}
\description{
Given atomic vector x find all indices i where x[i] == x[i + 1]
}
\examples{
neighboreq(rep(0, 10))
neighboreq(1:10)
}

