% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_prcomp.R
\name{plot_prcomp}
\alias{plot_prcomp}
\title{Plot first three principal components}
\usage{
plot_prcomp(prcomp, group, group_name = "group", main = "")
}
\arguments{
\item{prcomp}{prcomp}

\item{group}{vector of the group/class labels}

\item{group_name}{this will be used for the legend}

\item{main}{plot title}
}
\value{
ggplot
}
\description{
Plot first three principal components
}

