% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pubchem_drug_and_medication_information.R
\name{pubchem_drug_and_medication_information}
\alias{pubchem_drug_and_medication_information}
\title{Download Drug and Medication Information for a given compound from PubChem}
\usage{
pubchem_drug_and_medication_information(cid)
}
\arguments{
\item{cid}{}
}
\value{
list of data.frames
}
\description{
Download Drug and Medication Information for a given compound from PubChem
}

