% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile}
\alias{quantile}
\title{Compute quantiles corresponding to the given probabilities}
\usage{
quantile(x, p)
}
\arguments{
\item{x}{numeric vector}

\item{p}{numeric vector of probabilites}
}
\value{
quantiles of x
}
\description{
Compute quantiles corresponding to the given probabilities
}
\examples{
quantile(rnorm(100), 0.5)
quantile(1:100, c(0.25, 0.5, 0.75))
}

