% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randperm}
\alias{randperm}
\title{Generate a random permutation of a given numeric vector.}
\usage{
randperm(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
permuted x
}
\description{
Generate a random permutation of a given numeric vector.
}
\examples{
randperm(1:10)
letters[randperm(1:length(letters))]
randperm(iris$Species)
}

