% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_and_reencode.R
\name{read_and_reencode}
\alias{read_and_reencode}
\title{Read file in a binary mode and re-encode
Adapted from Marek Gagolewski, Advanced Data Analysis Software Development with R}
\usage{
read_and_reencode(fname, from = "utf-8", to = "",
  newlines = "\\\\r\\\\n|\\\\r|\\\\n")
}
\arguments{
\item{fname}{path to the input file}

\item{from}{input encoding}

\item{to}{output encoding}

\item{newlines}{character, regular expression}
}
\value{
character
}
\description{
Read file in a binary mode and re-encode
Adapted from Marek Gagolewski, Advanced Data Analysis Software Development with R
}

