% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replace_emails.R
\name{replace_emails}
\alias{replace_emails}
\title{Replace email addresses formed according to W3C recommendation}
\usage{
replace_emails(x, replacement = "__EMAIL__")
}
\arguments{
\item{x}{character vector}

\item{replacement}{character}
}
\value{
character vector
}
\description{
Replace email addresses formed according to W3C recommendation
}

