% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/replacena.R
\name{replacena}
\alias{replacena}
\title{Replace missing values in numeric vector
with arithmetic mean of its flanking elements
or throws an error if it is impossible.}
\usage{
replacena(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector
}
\description{
Replace missing values in numeric vector
with arithmetic mean of its flanking elements
or throws an error if it is impossible.
}
\examples{
replacena(c(1, NA, 2))
replacena(c(0, NA, 0, NA, 0))
}

