% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row_qvalues.R
\name{row_qvalues}
\alias{row_qvalues}
\title{Helper to compute rowwise qvalues}
\usage{
row_qvalues(x, fac)
}
\arguments{
\item{x}{numeric matrix. See genefilter::rowttests for details}

\item{fac}{group factor. See genefilter::rowttests for details}
}
\value{
qvalue object. See qvalue::qvalue for details
}
\description{
Helper to compute rowwise qvalues
}

