% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample2}
\alias{sample2}
\title{Generate a random, wihtout replacement,
subvector of length k of a given numeric vector x}
\usage{
sample2(x, k)
}
\arguments{
\item{x}{numeric vector}

\item{k}{number of elements}
}
\value{
random subvector of x
}
\description{
Generate a random, wihtout replacement,
subvector of length k of a given numeric vector x
}
\examples{
sample2(1:100, 5)
sample2(runif(5, 0, 100), 3)
sample2(iris$Sepal.Length, 20)
}

