% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scrabble.R
\name{scrabble}
\alias{scrabble}
\title{Return n most valuable words from dictionary}
\usage{
scrabble(x, n, dict, y = NULL)
}
\arguments{
\item{x}{named integer vector with values of individual letters}

\item{n}{number of words to return}

\item{dict}{legal words}

\item{y}{available letters}
}
\value{
character vector of length <= n
}
\description{
Return n most valuable words from dictionary
}
\examples{
scrabble(x = c(a = 1, b = 1, f = 1, o = 1, r = 1), 1, c("foo", "bar"))
scrabble(
   x = c(a = 1, b = 1, f = 1, o = 1, r = 1), 1,
   c("foo", "bar"),
   y = c(a = 1, b = 1, f = 0, o = 1, r = 1)
)
}

