% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortestpath}
\alias{shortestpath}
\title{Given binary adjacency matrix return shortest path between vertices}
\usage{
shortestpath(G, x, y)
}
\arguments{
\item{G}{integer matrix}

\item{x}{integer source}

\item{y}{integer sink}
}
\value{
shortest path integer
}
\description{
Given binary adjacency matrix return shortest path between vertices
}

