% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simplify2array}
\alias{simplify2array}
\title{Take list of integer vectors and convert to array}
\usage{
simplify2array(x)
}
\arguments{
\item{x}{List}
}
\value{
array
}
\description{
Take list of integer vectors and convert to array
}

