% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sortedmerge}
\alias{sortedmerge}
\title{Merge two sorted (nondecreasingly or nonincreasingly) vectors}
\usage{
sortedmerge(first, second)
}
\arguments{
\item{first}{numeric vector}

\item{second}{numeric vector}
}
\value{
numeric vector
}
\description{
Merge two sorted (nondecreasingly or nonincreasingly) vectors
}
\examples{
sortedmerge(c(1, 3, 5), c(2, 4, 6))
sortedmerge(3:1, 3:1)
}

