% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{Substitutes tags delimited by percent sign}
\usage{
template(dirname, data)
}
\arguments{
\item{dirname}{- directory containing ".tpl" files}

\item{data}{- data frame with two - columns key and value}
}
\description{
Substitutes tags delimited by percent sign
}

