% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/uncomment.R
\name{uncomment}
\alias{uncomment}
\title{Remove comments from a source code using tags list
It should be done with a proper lexer but who cares ;)}
\usage{
uncomment(text, tags)
}
\arguments{
\item{text}{character vector with input lines}

\item{tags}{2-column character matrix with comment strings. Use NA to indicate "\n" as closing tag}
}
\value{
text without comments and empty lines. More or less.
}
\description{
Remove comments from a source code using tags list
It should be done with a proper lexer but who cares ;)
}

