% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/unwind.R
\name{unwind}
\alias{unwind}
\title{Convert matrix from wide form to long}
\usage{
unwind(x, names)
}
\arguments{
\item{x}{matrix}

\item{names}{vector of names}
}
\value{
matrix in a long format
}
\description{
Convert matrix from wide form to long
}
\examples{
unwind(WorldPhones, c("count", "where", "when"))
unwind(
   matrix(c(1, 0, 0, 1), nrow=2, dimnames = list(c("x", "y"), c("x", "y"))),
   c("value", "c1", "c2")
)
}

