% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wilcoxteststat.R
\name{wilcoxteststat}
\alias{wilcoxteststat}
\title{Calculate Wilcoxon signed rank test}
\usage{
wilcoxteststat(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
Wilcoxon signed rank test statistic
}
\description{
Calculate Wilcoxon signed rank test
}
\examples{
wilcoxteststat(runif(10) + 1, rnorm(10))
wilcoxteststat(rnorm(100), rnorm(100))
wilcoxteststat(1:10, 1:10)
}

