% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/winsor.R
\name{winsor}
\alias{winsor}
\title{Calculate k-Winsorized mean of vector x.}
\usage{
winsor(x, k)
}
\arguments{
\item{x}{numeric vector of length n}

\item{k}{non-negative integer <= (n - 1) / 2}
}
\value{
k-Winsorized mean of the x
}
\description{
Calculate k-Winsorized mean of vector x.
}
\examples{
winsor(1:1000, 3)
winsor(rnorm(1000), 25)
winsor(runif(1000), 100)
}

