% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngram_tokenizer.R
\name{ngram_tokenizer}
\alias{ngram_tokenizer}
\title{Ngrams tokenizer}
\usage{
ngram_tokenizer(n = 1L, skip_word_none = TRUE)
}
\arguments{
\item{n}{integer}
}
\value{
n-gram tokenizer function
}
\description{
Ngrams tokenizer
}

